work

+++ src/idleRange.py
@@ -6,53 +6,56 @@ DB = 'gps_can'
 con = pg.connect(dbname=DB, host='localhost', user=USER,passwd='F1ff')
 
 test= False
-PREFIX = 'a'
-if len(sys.argv) > 1:
-	PREFIX = sys.argv[1]
-DATATABLE = ""+PREFIX+"_gps_can_data"
-TRIPDATA = ""+PREFIX+"_trip_data"
+#PREFIX = 'a'
+#if len(sys.argv) > 1:
+#	PREFIX = sys.argv[1]
+#DATATABLE = ""+PREFIX+"_gps_can_data"
+#TRIPDATA = ""+PREFIX+"_trip_data"
 
-"""
-print "Altering table"
-con.query('alter table ' + DATATABLE + ' drop IF EXISTS idleRange;')
-con.query('alter table ' + DATATABLE + ' add column idleRange int not null default 0;')
+try:
+	DATATABLE = sys.argv[1]
+except:
+	print 'Error: remember the parameters'
+	exit(1)
 
-con.query('update ' + DATATABLE + ' set idleRange = 1 where rpm>0 and speed = 0;')
+if False:
+	print "Altering table"
+	con.query('alter table ' + DATATABLE + ' drop IF EXISTS idleRange;')
+	con.query('alter table ' + DATATABLE + ' add column idleRange int not null default 0;')
 
-print "Creating index"
-con.query("DROP INDEX IF EXISTS idle_" + DATATABLE + "_idx;")
-con.query("Create index idle_" + DATATABLE + "_idx on " + DATATABLE + " (idleRange);")
+	con.query('update ' + DATATABLE + ' set idleRange = 1 where rpm>0 and speed = 0;')
 
-"""
+	print "Creating index"
+	con.query("DROP INDEX IF EXISTS idle_" + DATATABLE + "_idx;")
+	con.query("Create index idle_" + DATATABLE + "_idx on " + DATATABLE + " (idleRange);")
 
-con.query("drop table if exists vehicleIdleRange;")
-con.query("create table vehicleIdleRange (vehicleid bigint, idleRange int, fuel float);")
+
+con.query("drop table if exists "+DATATABLE+"_vehicleData;")
+con.query("create table "+DATATABLE+"_vehicleData (vehicleid bigint, idleRange int, fuel float, starTime timestamp);")
 print "Calculating"
 vehicles = con.query("select distinct vehicleid from "+ DATATABLE + ";").getresult()
 
 for v in vehicles:
-	res = con.query("select timestamp, idleRange, totalconsumed from " + DATATABLE + " where vehicleid=" + str(v[0]) +" order by timestamp;").getresult()
+	res = con.query("select timestamp, idleRange, totalconsumed from " + DATATABLE + " where vehicleid=" + str(v[0]) +" and dirty is false order by timestamp;").getresult()
 	s = -1
 	fuel = 0
+#TODO: timediff is too large, and current is one -> update
 	for r in range(0, len(res)-1):
 		if res[r][1] == 1 and s== -1:
 			s = time.mktime(time.strptime(res[r][0], "%Y-%m-%j %H:%M:%S"))
-#			print res[r][0]
+			startTime = res[r][0]
 			fuel = res[r][2]
-		if r > 0:
+		if r > 0 and res[r-1][1]==1:
 			timeDiff = abs(time.mktime(time.strptime(res[r-1][0], "%Y-%m-%j %H:%M:%S")) - time.mktime(time.strptime(res[r][0], "%Y-%m-%j %H:%M:%S")))
 		else:
 			timeDiff = 0
-		if (res[r][1] == 0 or ((r==len(res)-1 or timeDiff > 90) and res[r][1] == 1)) and not s==-1 and r > 0:
+		if (res[r][1] == 0 or ((r==len(res)-1 or timeDiff > 120) and res[r][1] == 1)) and not s==-1 and r > 0:
 			t = time.mktime(time.strptime(res[r-1][0], "%Y-%m-%j %H:%M:%S"))
 			if abs(t-s)> 0:
-#				if abs(t-s)>50000:
-#					print str(r) + "\t" + str(res[r][1]) + "\t" + str(timeDiff) +"\t" + str(s) + "\t" + str(t) + "\t" + str(abs(t-s)) + "\t" + res[r-1][0] + "\t" + str(v[0])
-#					break
 				fuel2 = res[r-1][2]
-				if fuel2 > 4294000:
-					fuel2 = res[r-2][2]
-				con.query("insert into vehicleIdleRange values (" + str(v[0]) + ", " + str(abs(t-s)) + "," + str(abs(fuel2-fuel)) + ")")
+				if startTime == '2012-09-03 06:18:55':
+					print str(v[0]) + '\t' + str(abs(t-s)) + '\t' + str(startTime) + '\t' + str(res[r-1][0])+ '\t' + str(fuel2) + '\t' + str(fuel) + '\t'
+				con.query("insert into "+DATATABLE+"_vehicleData values (" + str(v[0]) + ", " + str(abs(t-s)) + "," + str(abs(fuel2-fuel)) + ",'" + str(startTime) + "')")
 			s = -1
 			fuel = 0
